requirements.txt
django
djangorestframework
drf-yasg
django-cors-headers
filters-django
django-filter
PyQt5